refactor: extract chat interface functionality to ChatInterfaceSystem
Move chat UI management from app.js to dedicated ChatInterfaceSystem following ECS pattern:
- Extract initChatInterface, sendMessage, addMessage, autoResizeTextarea, updateSendButton methods
- Add handleImageUpload and handlePaste for complete image support (upload + paste)
- Maintain all existing functionality including image uploads, model selection, and commands
- Add helper methods in app.js to delegate to ChatInterfaceSystem
- Reduce app.js size by 260 lines while adhering to ECS architecture
