+++ src/js/app.js
@@ -14,6 +14,7 @@ import {
   PersistenceSystem,
   VoxelIndicatorRenderSystem,
   DOMInterfaceSystem,
+  ChatInterfaceSystem,
 } from "./systems/index.js";
 import {
   Connection,
@@ -40,13 +41,35 @@ class IndustrialPortfolio {
 
     try {
       // Theme, Font, Navigation, and Sidebar now handled by DOMInterfaceSystem
-      this.initChatInterface();
+      // Chat interface now handled by ChatInterfaceSystem
       this.initSessionsList();
       await this.initECS();
       this.initialized = true;
       console.log("✅ Industrial Portfolio initialized successfully");
     } catch (error) {
       console.error("❌ Error initializing portfolio:", error);
+    }
+  }
+
+  // Helper methods to delegate to ChatInterfaceSystem
+  addMessage(type, content) {
+    const chatInterface = this.world?.getSystem("chatInterface");
+    if (chatInterface) {
+      chatInterface.addMessage(type, content);
+    }
+  }
+
+  autoResizeTextarea(textarea) {
+    const chatInterface = this.world?.getSystem("chatInterface");
+    if (chatInterface) {
+      chatInterface.autoResizeTextarea(textarea);
+    }
+  }
+
+  updateSendButton(input, button) {
+    const chatInterface = this.world?.getSystem("chatInterface");
+    if (chatInterface) {
+      chatInterface.updateSendButton(input, button);
     }
   }
 
@@ -57,6 +80,11 @@ class IndustrialPortfolio {
     const domInterface = new DOMInterfaceSystem();
     this.world.addSystem(domInterface, "domInterface");
     domInterface.init();
+
+    // Add Chat Interface system for chat UI management
+    const chatInterface = new ChatInterfaceSystem();
+    this.world.addSystem(chatInterface, "chatInterface");
+    chatInterface.init(this.world, this);
 
     // Add core systems
     this.world.addSystem(new RenderSystem(), "render");
@@ -249,74 +277,7 @@ class IndustrialPortfolio {
     indicator.createPattern("idle");
   }
 
-  initChatInterface() {
-    console.log("💬 Initializing chat interface...");
-
-    const chatInput = document.getElementById("chat-input");
-    const chatSend = document.getElementById("chat-send");
-    const modelSelect = document.getElementById("model-select");
-    const imageUploadBtn = document.getElementById("image-upload-btn");
-    const imageUpload = document.getElementById("image-upload");
-
-    if (chatInput && chatSend) {
-      // Auto-resize textarea
-      chatInput.addEventListener("input", () => {
-        this.autoResizeTextarea(chatInput);
-        this.updateSendButton(chatInput, chatSend);
-      });
-
-      // Handle Enter key
-      chatInput.addEventListener("keydown", (e) => {
-        if (e.key === "Enter" && !e.shiftKey) {
-          e.preventDefault();
-          this.sendMessage();
-        }
-      });
-
-      // Handle send button
-      chatSend.addEventListener("click", () => {
-        this.sendMessage();
-      });
-
-      // Initial state
-      this.updateSendButton(chatInput, chatSend);
-
-      // Handle focus event to activate session
-      chatInput.addEventListener("focus", () => {
-        this.activatePlayerOriginSession();
-      });
-    }
-
-    // Model selector
-    if (modelSelect) {
-      modelSelect.addEventListener("change", (e) => {
-        const agentSystem = this.world.getSystem("agent");
-        if (agentSystem) {
-          agentSystem.switchModel(e.target.value);
-        }
-      });
-    }
-
-    // Image upload
-    if (imageUploadBtn && imageUpload) {
-      imageUploadBtn.addEventListener("click", () => {
-        imageUpload.click();
-      });
-
-      imageUpload.addEventListener("change", async (e) => {
-        await this.handleImageUpload(e.target.files);
-      });
-    }
-
-    // Clipboard paste support
-    if (chatInput) {
-      chatInput.addEventListener("paste", async (e) => {
-        await this.handlePaste(e);
-      });
-    }
-
-    console.log("✅ Chat interface initialized");
-  }
+  // Chat interface now handled by ChatInterfaceSystem
 
   initSessionsList() {
     console.log("💬 Initializing sessions list...");
@@ -1599,121 +1560,9 @@ class IndustrialPortfolio {
 
 
 
-  autoResizeTextarea(textarea) {
-    textarea.style.height = "auto";
-    textarea.style.height = Math.min(textarea.scrollHeight, 200) + "px";
-  }
+  // autoResizeTextarea and updateSendButton now handled by ChatInterfaceSystem
 
-  updateSendButton(input, button) {
-    const hasText = input.value.trim().length > 0;
-    button.disabled = !hasText;
-  }
-
-  async sendMessage() {
-    const chatInput = document.getElementById("chat-input");
-    const message = chatInput.value.trim();
-
-    if (!message) return;
-
-    console.log("📤 Sending message:", message);
-
-    // Check for slash commands
-    if (message.startsWith("/")) {
-      await this.handleSlashCommand(message);
-      return;
-    }
-
-    // Ensure session is active with current chat target
-    const session = this.activatePlayerTargetSession();
-    if (!session) {
-      this.addMessage(
-        "assistant",
-        "Failed to establish connection with target entity.",
-      );
-      return;
-    }
-
-    // Get images if any
-    const imagePreview = document.getElementById("image-preview");
-    const images = [];
-    if (imagePreview) {
-      const imageElements = imagePreview.querySelectorAll("img");
-      for (const img of imageElements) {
-        const base64 = img.getAttribute("data-base64");
-        if (base64) images.push(base64);
-      }
-    }
-
-    // Clear input and images
-    chatInput.value = "";
-    this.autoResizeTextarea(chatInput);
-    this.updateSendButton(chatInput, document.getElementById("chat-send"));
-    if (imagePreview) {
-      imagePreview.innerHTML = "";
-      // Hide image container
-      const imageContainer = document.getElementById("image-upload-container");
-      if (imageContainer) {
-        imageContainer.style.display = "none";
-      }
-    }
-
-    // Send message through session system
-    const sessionSystem = this.world.getSystem("session");
-    const agentSystem = this.world.getSystem("agent");
-
-    if (sessionSystem && this.playerEntity && this.currentChatTarget) {
-      // Add user message to session with images
-      sessionSystem.sendMessage(
-        session.id,
-        this.playerEntity.id,
-        message,
-        "user",
-        { images: images }
-      );
-
-      // Display user message
-      this.addMessage("user", message);
-
-      // Get AI response if agent system is ready
-      if (agentSystem && agentSystem.isConnected) {
-        // Get response from AI using entity context
-        const response = await agentSystem.generateResponseWithContext(
-          message,
-          this.currentChatTarget,
-          { images: images, userMessage: message },
-        );
-
-        // Add AI response to session
-        sessionSystem.sendMessage(
-          session.id,
-          this.currentChatTarget.id,
-          response,
-          "llm",
-        );
-
-        // Display AI response
-        this.addMessage("assistant", response);
-        
-        // Refresh sessions list to show updated activity
-        this.loadSessionsList();
-      } else {
-        // Fallback response
-        const fallbackResponse =
-          "Agent system not initialized. Please check if Ollama is running.";
-        sessionSystem.sendMessage(
-          session.id,
-          this.currentChatTarget.id,
-          fallbackResponse,
-          "system",
-        );
-        this.addMessage("assistant", fallbackResponse);
-      }
-    } else {
-      // Fallback if session system not ready
-      this.addMessage("user", message);
-      this.addMessage("assistant", "Session system not initialized.");
-    }
-  }
+  // sendMessage now handled by ChatInterfaceSystem
 
   async handleSlashCommand(command) {
     const chatInput = document.getElementById("chat-input");
@@ -1770,120 +1619,11 @@ class IndustrialPortfolio {
     }
   }
 
-  addMessage(type, content) {
-    const chatMessages = document.getElementById("chat-messages");
-    const welcome = chatMessages.querySelector(".chat-welcome");
+  // addMessage now handled by ChatInterfaceSystem via helper method above
 
-    // Remove welcome message if it exists
-    if (welcome) {
-      welcome.remove();
-    }
+  // handleImageUpload now handled by ChatInterfaceSystem
 
-    // Create message element
-    const messageDiv = document.createElement("div");
-    messageDiv.className = `message ${type}`;
-
-    const messageContent = document.createElement("div");
-    messageContent.className = "message-content";
-    messageContent.textContent = content;
-
-    messageDiv.appendChild(messageContent);
-    chatMessages.appendChild(messageDiv);
-
-    // Scroll to bottom
-    chatMessages.scrollTop = chatMessages.scrollHeight;
-  }
-
-  async handleImageUpload(files) {
-    const imagePreview = document.getElementById("image-preview");
-    const agentSystem = this.world.getSystem("agent");
-
-    if (!agentSystem) {
-      alert("AgentSystem not ready. Please wait for Ollama to connect.");
-      return;
-    }
-
-    if (!agentSystem.supportsImages) {
-      alert(
-        `Current model "${agentSystem.currentModel}" doesn't support images. Please select a multimodal model like gemma2 or gemma3.`,
-      );
-      return;
-    }
-
-    // Show image upload container if hidden
-    const imageContainer = document.getElementById("image-upload-container");
-    if (imageContainer && files.length > 0) {
-      imageContainer.style.display = "block";
-    }
-
-    // Process each file
-    for (const file of files) {
-      if (!file.type.startsWith("image/")) continue;
-
-      try {
-        const base64 = await agentSystem.processImage(file);
-
-        // Create preview
-        const img = document.createElement("img");
-        img.src = `data:${file.type};base64,${base64}`;
-        img.setAttribute("data-base64", base64);
-        img.title = "Click to remove";
-        img.addEventListener("click", () => {
-          img.remove();
-          // Hide container if no images left
-          if (imagePreview.children.length === 0) {
-            const imageContainer = document.getElementById(
-              "image-upload-container",
-            );
-            if (imageContainer) {
-              imageContainer.style.display = "none";
-            }
-          }
-        });
-
-        imagePreview.appendChild(img);
-      } catch (error) {
-        console.error("Failed to process image:", error);
-      }
-    }
-  }
-
-  async handlePaste(event) {
-    const clipboardData = event.clipboardData || window.clipboardData;
-    if (!clipboardData) {
-      console.log("📋 No clipboard data available");
-      return;
-    }
-
-    const items = clipboardData.items;
-    const imageFiles = [];
-
-    console.log("📋 Clipboard items:", items.length);
-    for (let i = 0; i < items.length; i++) {
-      console.log(`📋 Item ${i}: type=${items[i].type}, kind=${items[i].kind}`);
-    }
-
-    // Look for images in clipboard
-    for (let i = 0; i < items.length; i++) {
-      const item = items[i];
-      if (item.type.startsWith("image/")) {
-        event.preventDefault(); // Prevent default paste behavior for images
-        const file = item.getAsFile();
-        if (file) {
-          imageFiles.push(file);
-          console.log("📋 Found image file:", file.name, file.type, file.size);
-        }
-      }
-    }
-
-    // Process images if found
-    if (imageFiles.length > 0) {
-      console.log("📋 Pasted", imageFiles.length, "image(s)");
-      await this.handleImageUpload(imageFiles);
-    } else {
-      console.log("📋 No images found in clipboard");
-    }
-  }
+  // handlePaste now handled by ChatInterfaceSystem
 
   // Debug method to access input system
   getInputSystem() {
