+++ src/js/systems/ChatInterface/index.js
@@ -1,0 +1,1 @@
+export { ChatInterfaceSystem } from './System.js';
