+++ src/js/systems/index.js
@@ -12,3 +12,4 @@ export { SessionSystem } from './Session/index.js';
 export { PersistenceSystem } from './Persistence/index.js';
 export { VoxelIndicatorRenderSystem } from './VoxelIndicatorRender/index.js';
 export { DOMInterfaceSystem } from './DOMInterface/index.js';
+export { ChatInterfaceSystem } from './ChatInterface/index.js';
